backend/requirements-train-noytorch.txt
# Training requirements WITHOUT PyTorch (installed separately for RTX 5070 compatibility)
# PyTorch nightly is installed in Dockerfile for sm_120 support

# Model architecture
timm>=0.9.12

# Metrics and evaluation
torchmetrics>=1.0.0
scikit-learn>=1.3.0

# Face detection
facenet-pytorch>=2.5.3

# Data augmentation
albumentations>=1.3.1

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Progress and logging
tqdm>=4.65.0
tensorboard>=2.14.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Video processing
decord>=0.6.0
av>=10.0.0

# Image processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# File downloads
gdown>=4.7.1
requests>=2.31.0

# Hugging Face Hub for model weights
huggingface_hub>=0.16.0
safetensors>=0.3.0

# Utilities from main requirements
aiofiles>=23.1.0
aiosqlite>=0.19.0
alembic>=1.11.0
asyncpg>=0.28.0
bcrypt>=4.0.0
celery>=5.3.0
cryptography>=41.0.0
fastapi>=0.100.0
httpx>=0.24.0
passlib>=1.7.4
prometheus-client>=0.17.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
redis>=4.6.0
reportlab>=4.0.0
SQLAlchemy>=2.0.0
uvicorn>=0.23.0
matplotlib>=3.7.0
